refactor ranking pipeline to statically ensure we score the different stages as expected
