+++ crates/core/examples/search_preindexed.rs
@@ -9,7 +9,7 @@ use stract::{
     image_store::Image,
     index::Index,
     inverted_index::RetrievedWebpage,
-    ranking::{inbound_similarity::InboundSimilarity, pipeline::RetrievedWebpageRanking},
+    ranking::{inbound_similarity::InboundSimilarity, pipeline::PrecisionRankingWebpage},
     searcher::{api::ApiSearcher, live::LiveSearcher, LocalSearcher, SearchQuery, ShardId},
     Result,
 };
@@ -30,9 +30,9 @@ impl stract::searcher::distributed::SearchClient for Searcher {
 
     async fn retrieve_webpages(
         &self,
-        top_websites: &[(usize, stract::searcher::ScoredWebsitePointer)],
+        top_websites: &[(usize, stract::searcher::ScoredWebpagePointer)],
         query: &str,
-    ) -> Vec<(usize, stract::ranking::pipeline::RetrievedWebpageRanking)> {
+    ) -> Vec<(usize, stract::ranking::pipeline::PrecisionRankingWebpage)> {
         let pointers = top_websites
             .iter()
             .map(|(_, p)| p.website.pointer.clone())
@@ -44,7 +44,7 @@ impl stract::searcher::distributed::SearchClient for Searcher {
             .unwrap()
             .into_iter()
             .zip(top_websites.iter().map(|(i, p)| (*i, p.website.clone())))
-            .map(|(ret, (i, ran))| (i, RetrievedWebpageRanking::new(ret, ran)))
+            .map(|(ret, (i, ran))| (i, PrecisionRankingWebpage::new(ret, ran)))
             .collect::<Vec<_>>();
 
         res
