+++ crates/core/src/searcher/distributed.rs
@@ -29,8 +29,8 @@ use crate::{
         search_server::{self, SearchService},
     },
     image_store::Image,
-    inverted_index::{RetrievedWebpage, WebsitePointer},
-    ranking::pipeline::{RankingWebsite, RetrievedWebpageRanking},
+    inverted_index::{RetrievedWebpage, WebpagePointer},
+    ranking::pipeline::{PrecisionRankingWebpage, RecallRankingWebpage},
     Result,
 };
 
@@ -59,8 +59,8 @@ pub enum Error {
 }
 
 #[derive(Clone, Debug)]
-pub struct ScoredWebsitePointer {
-    pub website: RankingWebsite,
+pub struct ScoredWebpagePointer {
+    pub website: RecallRankingWebpage,
     pub shard: ShardId,
 }
 
@@ -126,9 +126,9 @@ impl DistributedSearcher {
         shard: ShardId,
         client: &ShardedClient<SearchService, ShardId>,
         query: &str,
-        pointers: Vec<(usize, WebsitePointer)>,
+        pointers: Vec<(usize, WebpagePointer)>,
     ) -> Vec<(usize, RetrievedWebpage)> {
-        let (idxs, pointers): (Vec<usize>, Vec<WebsitePointer>) = pointers.into_iter().unzip();
+        let (idxs, pointers): (Vec<usize>, Vec<WebpagePointer>) = pointers.into_iter().unzip();
 
         match client
             .send(
@@ -184,9 +184,9 @@ impl SearchClient for DistributedSearcher {
 
     async fn retrieve_webpages(
         &self,
-        top_websites: &[(usize, ScoredWebsitePointer)],
+        top_websites: &[(usize, ScoredWebpagePointer)],
         query: &str,
-    ) -> Vec<(usize, RetrievedWebpageRanking)> {
+    ) -> Vec<(usize, PrecisionRankingWebpage)> {
         let mut rankings = FnvHashMap::default();
         let mut pointers: HashMap<_, Vec<_>> = HashMap::new();
 
@@ -209,7 +209,7 @@ impl SearchClient for DistributedSearcher {
         for pages in join_all(futures).await {
             for (i, page) in pages {
                 retrieved_webpages
-                    .push((i, RetrievedWebpageRanking::new(page, rankings[&i].clone())));
+                    .push((i, PrecisionRankingWebpage::new(page, rankings[&i].clone())));
             }
         }
 
@@ -308,9 +308,9 @@ pub trait SearchClient {
 
     fn retrieve_webpages(
         &self,
-        top_websites: &[(usize, ScoredWebsitePointer)],
+        top_websites: &[(usize, ScoredWebpagePointer)],
         query: &str,
-    ) -> impl Future<Output = Vec<(usize, RetrievedWebpageRanking)>> + Send;
+    ) -> impl Future<Output = Vec<(usize, PrecisionRankingWebpage)>> + Send;
 
     fn search_entity(&self, query: &str) -> impl Future<Output = Option<EntityMatch>> + Send;
